# perf(Bring back twitter): Optimize string replacement