--- a/src/main/kotlin/crimera/patches/twitter/misc/bringbacktwitter/BringBackTwitterResourcePatch.kt
+++ b/src/main/kotlin/crimera/patches/twitter/misc/bringbacktwitter/BringBackTwitterResourcePatch.kt
@@ -119,24 +119,27 @@ object BringBackTwitterResourcePatch : ResourcePatch() {
             for (i in 0 until nodes.length) {
                 val node = nodes.item(i)
                 val name = node.attributes.getNamedItem("name")?.nodeValue
+
                 if (name == "conference_default_title") {
                     /*
-                     * Parsing causes the default value which contains the
-                     * character 𝕏 to be "corrupted" so we change it to a normal X
+                     * Parsing causes the string which contains the character "𝕏" to be corrupted,
+                     * so we change it to "Twitter"
                      */
-                    val content = node.textContent
                     val value = stringsMap[name]
                     if (value != null) {
                         node.textContent = value
                         replacedCount++
                     } else {
+                        val content = node.textContent
                         val delimiter = if (content.contains("-")) '-' else ' '
-                        node.textContent = content.split(delimiter).joinToString(delimiter.toString()) {
-                            if (it.startsWithSpecialByte()) "Twitter" else it
-                        }
+                        node.textContent =
+                            content.split(delimiter).joinToString(delimiter.toString()) {
+                                if (it.startsWithSpecialByte()) "Twitter" else it
+                            }
                     }
                     continue
                 }
+
                 node.textContent = stringsMap[name] ?: continue
                 replacedCount++
             }