https://t.me/pikopatches/1/17451

The current implementation has poor performance as it iterates over all string elements for each entry in the stringsMap.
DOM operations using Java XML API are expensive, but on the other hand, Map lookups are fast.
So, the string replacement should be done in a single iteration.

I tested performance on a very low-end phone (Snapdragon 410, 2GB RAM) with ReVanced Manager.

Before PR:

```
values: 2373 ms
values-en-rGB: 1886 ms
values-ru: 1983 ms
values-hi: 1986 ms
values-pt-rBR: 17 ms
values-tr: 1946 ms
values-zh-rCN: 1950 ms
values-zh-rTW: 2008 ms
Total: 14149 ms
```

After PR:

```
values: 631 ms
values-en-rGB: 339 ms
values-ru: 445 ms
values-hi: 453 ms
values-pt-rBR: 7 ms
values-tr: 367 ms
values-zh-rCN: 400 ms
values-zh-rTW: 392 ms
Total: 3034 ms
```

**[ADDED]**
Furthermore, with this change, adding more strings to the map will have little impact on the performance of the patch.
Thus, more "X" can be replaced with "Twiter".

---

I have a few questions before this PR is ready.

- Do you really need to add strings that aren't present in the original strings.xml?

If a string does not exist in the original strings.xml, it was probably deleted from the app.
If so, I think adding it is no point because it will never displayed in the app.
In addition I can remove `mutableStringsMap`.

- Do you really need to create a new string.xml when it doesn't exist?

In my opinion, it is not needed to create a new strings.xml even if a specific language string.xml is not found, 
When users merged the split APKs, they may have removed unnecessary language splits, in which case this patch does not need to add strings.xml for languages ​​that do not exist.
Also, if a user unintentionally excludes the language split for their language, after applying this patch they will experience a strange situation where only the 60 strings included in this patch are translated.

At the time this patch was created, full APK of X was provided.
However, now that users have to merge the split APKs themselves, I believe that this behavior needs to be reconsidered.

- What is `// log which keys were not found or failed` ?

I can't get what this comment means.
I think logging the missing strings is good idea instead of adding them silently.