@crimera 2024-10-20
> What is // log which keys were not found or failed ?
I think this is for testing that we forgot to remove.
> Do you really need to create a new string.xml when it doesn't exist?
> Do you really need to add strings that aren't present in the original strings.xml?
Fair points. Now that I look at the code, they really are redundant. One counter point. It may be that the original apk does not provide translations for the user's native language, so they create a custom translation for it. But seeing that no one actually did that, then I think it is safe to remove them after all.