@kitadai31 2024-10-20
Hey, after I submitted this PR, I found a old code in the git log that replaces all "X" with "Twitter", and I was curious why it was slow, so I looked into it.
https://github.com/crimera/piko/blob/2c63ff1bc1444ed6d83f911becca4289001c306c/src/main/kotlin/crimera/patches/twitter/misc/bringbacktwitter/BringBackTwitterResourcePatch.kt

In conclusion, it was a problem with using setTextContent().

This part was causing the performance issue.
It called `setTextContent()` for each string. (= 6000-7000 times for each language.)

```kt
for (i in 0 until strings.length) {
    val string = strings.item(i) as Element
    if (!string.getAttribute("name").contains("api_key")) {
        string.textContent = string.textContent.replace("X", "Twitter")
    }
}
```

Kotlin makes setters like `setTextContent()` look like a simple assignment, but in reality `setTextContent()` is a very slow operation.
I tried rewriting this part like this:

```kt
for (i in 0 until strings.length) {
    val string = strings.item(i) as Element
    if (!string.getAttribute("name").contains("api_key")) {
        val text = string.textContent
        if (text.contains("X")) {
            string.textContent = text.replace("X", "Twitter")
        }
    }
}
```

This improved performance incredibly.

```
Old code:
res/values-de/strings.xml: 1325 ms
res/values-ko/strings.xml: 1272 ms
res/values-pt/strings.xml: 1220 ms
res/values-ro/strings.xml: 1256 ms
res/values-zh-rCN/strings.xml: 1231 ms
res/values-ar-rEH/strings.xml: 985 ms
res/values-hu/strings.xml: 1506 ms
res/values-fr/strings.xml: 1266 ms
res/values-ja/strings.xml: 1235 ms
res/values-uk/strings.xml: 1264 ms
res/values-it/strings.xml: 1236 ms
res/values-in/strings.xml: 1269 ms
res/values-pl/strings.xml: 1321 ms
res/values-hi/strings.xml: 1289 ms
res/values-ru/strings.xml: 1230 ms
res/values-ms/strings.xml: 1247 ms
res/values-th/strings.xml: 1214 ms
res/values-nl/strings.xml: 1293 ms
res/values-en-rGB/strings.xml: 1632 ms
res/values-zh-rTW/strings.xml: 1287 ms
res/values-zh-rHK/strings.xml: 1241 ms
res/values-es/strings.xml: 1297 ms
res/values-fi/strings.xml: 1250 ms
res/values-vi/strings.xml: 1299 ms
res/values-ar/strings.xml: 1303 ms
res/values-sv/strings.xml: 1257 ms
res/values-tr/strings.xml: 1290 ms
res/values/strings.xml: 2050 ms

↓↓↓↓↓↓ 

New code:
res/values-de/strings.xml: 128 ms
res/values-ko/strings.xml: 95 ms
res/values-pt/strings.xml: 70 ms
res/values-ro/strings.xml: 75 ms
res/values-zh-rCN/strings.xml: 74 ms
res/values-ar-rEH/strings.xml: 68 ms
res/values-hu/strings.xml: 70 ms
res/values-fr/strings.xml: 68 ms
res/values-ja/strings.xml: 57 ms
res/values-uk/strings.xml: 29 ms
res/values-it/strings.xml: 48 ms
res/values-in/strings.xml: 50 ms
res/values-pl/strings.xml: 48 ms
res/values-hi/strings.xml: 70 ms
res/values-ru/strings.xml: 61 ms
res/values-ms/strings.xml: 49 ms
res/values-th/strings.xml: 58 ms
res/values-nl/strings.xml: 51 ms
res/values-en-rGB/strings.xml: 50 ms
res/values-zh-rTW/strings.xml: 55 ms
res/values-zh-rHK/strings.xml: 57 ms
res/values-es/strings.xml: 51 ms
res/values-fi/strings.xml: 51 ms
res/values-vi/strings.xml: 99 ms
res/values-ar/strings.xml: 55 ms
res/values-sv/strings.xml: 51 ms
res/values-tr/strings.xml: 51 ms
res/values/strings.xml: 64 ms
```

> Please note that this result is measured on a Windows PC with Ryzen 5 5500U + 8GB RAM using ReVanced CLI.
So this time cannot be comparable with the time above.
I have not tried the old code on the low-end phone, but it is not hard to imagine that the results would be very bad.

This execution time is almost the same as the current process using replacement map.
With this change, performance is no longer an issue, so in the future we could consider simply replacing all X with Twitter.

But this is beyond the scope of this PR. I just tell this for yours information.