@kitadai31 2024-10-20
I pushed the change.
I left a log code but it shows the number of strings which was not found.
Please check if this is fine.

However, a new problem with `pt-rBR` languages arose.
This patch has a pt-rBR strings map, but the X APK only has `values-pt` and doesn't have `values-pt-rBR`.
pt-rBR strings will not be applied after this change.

I don't know the difference between normal pt and pt-rBR, but if the strings in the pt_rBR map is taken directly from Twitter 9.98.0 apk, I think I can change values-pt-rBR to values-pt directly.

I'll look into it tomorrow or the day after tomorrow.