@crimera 2024-10-20
> pt-rBR languages

It's for Portuguese Region Brazil. It seems I missed this one. We actually need to create the strings for that.