@crimera 2024-10-22
> Should I keep these code after merging conflicts?

Nah, I forgot to removs it.