@kitadai31 2024-11-10
Sorry for the late reply.
I pushed the change, and this PR is now ready to be reviewed or merged.

By the way, I removed the missing strings logs at https://github.com/crimera/piko/commit/3448546ea61d5f387b5569e59970bde838b0b111.

It turned out that when user removed some languages from split apk and SettingsPatch was applied prior to this patch, this log always show warnings because SettingsPatch adds other languages files.
For example, when a user chose only English split when merging split APKs, this log show a warning for each `en-rGB`, `ru`, `hi`, `tr`, `zh-rCN`, and `zh-rTW` languages that says "60 keys were not found in ..."

Therefore, I removed the log.

(If you still think the log is needed, please let me know.)