@kitadai31 2024-11-10
I forgot to tell you one important thing.
Sincerely, please merge with squash and merge.

When I resolved the conflict on my local repository, I should have rebased the working branch on the dev branch instead of merging the dev branch. 😅
(Many "Merge branch 'refs/heads/dev' into fix/bring-back-twitter-performance")
As a result, this commit log is dirty, so doing a normal merge will mess up the commit log. 😅